requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2
python-dotenv>=1.0
# API/взаимодействие с Google Gemini (новый SDK)
google-genai>=0.3.0
# Если будет REST API
djangorestframework>=3.15
# Если понадобится WebSocket (опционально)
channels>=4.0
channels-redis>=4.2
# Ограничение запросов
django-ratelimit>=4.1
# Отладка/мониторинг (по желанию)
sentry-sdk>=2.0
markdown2>=2.4

# For RAG (Retrieval-Augmented Generation)
langchain>=0.1.20
langchain-google-genai>=1.0.2
langchain-community>=0.0.36
faiss-cpu>=1.8.0
pypdf>=4.2.0
